move steam warning loading behind adapter hooks
