+++ src/lib/shared/useAccountLoader.svelte.ts
@@ -1,80 +1,10 @@
 import type { PlatformAdapter, PlatformAccount } from "./platform";
-import type { BanInfo } from "../platforms/steam/types";
-import { addToast, removeToast } from "../features/notifications/store.svelte";
-import { getPlayerBans, hasApiKey } from "../platforms/steam/steamApi";
+import { addToast } from "../features/notifications/store.svelte";
+import type { AccountWarningPresentation } from "./accountWarnings";
 import { DEFAULT_LOCALE, translate, type MessageKey, type TranslationParams } from "$lib/i18n";
 
-import { getSettings } from "../features/settings/store";
-
 const BATCH_SIZE = 5;
-const BAN_CHECK_STATE_KEY = "accshift_ban_check_state_v2";
-const BAN_INFO_CACHE_KEY = "accshift_ban_info_cache_v1";
-const BAN_ERROR_TOAST_COOLDOWN_MS = 30000;
 const LOAD_TOAST_COOLDOWN_MS = 30000;
-
-interface BanCheckState {
-  lastSuccessAt: number;
-  checkedSteamIds: string[];
-}
-
-function readBanCheckState(): BanCheckState | null {
-  try {
-    const raw = localStorage.getItem(BAN_CHECK_STATE_KEY);
-    if (!raw) return null;
-    const parsed = JSON.parse(raw) as Partial<BanCheckState>;
-    const lastSuccessAt = Number(parsed.lastSuccessAt);
-    const checkedSteamIds = Array.isArray(parsed.checkedSteamIds)
-      ? parsed.checkedSteamIds.filter((id): id is string => typeof id === "string")
-      : [];
-    if (!Number.isFinite(lastSuccessAt) || lastSuccessAt < 0) return null;
-    return {
-      lastSuccessAt,
-      checkedSteamIds: Array.from(new Set(checkedSteamIds)),
-    };
-  } catch {
-    return null;
-  }
-}
-
-function writeBanCheckState(state: BanCheckState) {
-  localStorage.setItem(BAN_CHECK_STATE_KEY, JSON.stringify(state));
-}
-
-function isBanInfo(value: unknown): value is BanInfo {
-  if (!value || typeof value !== "object") return false;
-  const v = value as Partial<BanInfo>;
-  return typeof v.steam_id === "string"
-    && typeof v.community_banned === "boolean"
-    && typeof v.vac_banned === "boolean"
-    && typeof v.number_of_vac_bans === "number"
-    && typeof v.days_since_last_ban === "number"
-    && typeof v.number_of_game_bans === "number"
-    && typeof v.economy_ban === "string";
-}
-
-function readBanInfoCache(): Record<string, BanInfo> {
-  try {
-    const raw = localStorage.getItem(BAN_INFO_CACHE_KEY);
-    if (!raw) return {};
-    const parsed = JSON.parse(raw) as Record<string, unknown>;
-    if (!parsed || typeof parsed !== "object") return {};
-
-    const entries = Object.entries(parsed).filter(([, value]) => isBanInfo(value));
-    return Object.fromEntries(entries.map(([key, value]) => [key, value as BanInfo]));
-  } catch {
-    return {};
-  }
-}
-
-function writeBanInfoCache(bans: Record<string, BanInfo>) {
-  localStorage.setItem(BAN_INFO_CACHE_KEY, JSON.stringify(bans));
-}
-
-function isSteamPathMissingError(message: string): boolean {
-  const normalized = message.trim().toLowerCase();
-  return normalized.includes("could not locate steam installation")
-    || normalized.includes("could not read steam configuration");
-}
 
 function deferBackgroundTask(task: () => void | Promise<void>) {
   if (typeof window !== "undefined" && "requestIdleCallback" in window) {
@@ -91,11 +21,10 @@ function deferBackgroundTask(task: () => void | Promise<void>) {
 
 export function createAccountLoader(
   getAdapter: () => PlatformAdapter | undefined,
-  getActiveTab: () => string,
   getVisibleAccountIds?: () => string[],
   translateMessage?: (key: MessageKey, params?: TranslationParams) => string,
 ) {
-  // Centralized UI state for account loading, switching, avatars, and Steam ban checks.
+  // Centralized UI state for account loading, switching, avatars, and platform warnings.
   let accounts = $state<PlatformAccount[]>([]);
   let accountMap = $derived<Record<string, PlatformAccount>>(
     Object.fromEntries(accounts.map(a => [a.id, a]))
@@ -105,10 +34,7 @@ export function createAccountLoader(
   let switching = $state(false);
   let error = $state<string | null>(null);
   let avatarStates = $state<Record<string, { url: string | null; loading: boolean; refreshing: boolean }>>({});
-  let banStates = $state<Record<string, BanInfo>>(readBanInfoCache());
-  let sessionBanCheckedIds = new Set<string>();
-  let lastBanErrorToastAt = 0;
-  let activeBanCheckToastId: string | null = null;
+  let warningStates = $state<Record<string, AccountWarningPresentation>>({});
   let lastSteamPathToastAt = 0;
   let lastNoAccountsToastAt = 0;
   let latestLoadId = 0;
@@ -188,144 +114,14 @@ export function createAccountLoader(
     }
   }
 
-  async function fetchBanStates(accts: PlatformAccount[], silent = false, forceRefresh = false) {
-    if (getActiveTab() !== "steam" || accts.length === 0) return;
-
-    const delayDays = getSettings().banCheckDays;
-    const forceBanRefresh = forceRefresh;
-    const steamIds = Array.from(new Set(accts.map(a => a.id)));
-    if (steamIds.length === 0) return;
-
-    const hasApiKeyConfigured = await hasApiKey().catch((e) => {
-      console.error("[ban-check] failed to detect API key availability:", e);
-      return false;
-    });
-    if (!hasApiKeyConfigured) {
-      console.info("[ban-check] skipped: missing Steam API key");
-      return;
-    }
-
-    const now = Date.now();
-    const cachedState = readBanCheckState();
-    const delayMs = delayDays * 24 * 60 * 60 * 1000;
-    const withinDelayWindow = delayDays > 0
-      && !!cachedState
-      && now - cachedState.lastSuccessAt < delayMs;
-    const cachedCheckedIds = new Set(cachedState?.checkedSteamIds ?? []);
-
-    // Refresh policy:
-    // - forceRefresh: always fetch all accounts
-    // - delayDays = 0: fetch once per session
-    // - delayDays > 0: fetch only unchecked IDs inside the delay window
-    let idsToFetch: string[] = [];
-    if (forceBanRefresh) {
-      idsToFetch = steamIds;
-    } else if (delayDays === 0) {
-      idsToFetch = steamIds.filter(id => !sessionBanCheckedIds.has(id));
-    } else if (withinDelayWindow) {
-      idsToFetch = steamIds.filter(id => !cachedCheckedIds.has(id));
-    } else {
-      idsToFetch = steamIds;
-    }
-    if (idsToFetch.length === 0) {
-      console.info("[ban-check] skipped: no accounts to check", {
-        totalAccounts: steamIds.length,
-        delayDays,
-        forceBanRefresh,
-      });
-      return;
-    }
-
-    let checkingToastId: string | null = null;
-    if (!silent) {
-      if (activeBanCheckToastId) {
-        removeToast(activeBanCheckToastId);
-      }
-      activeBanCheckToastId = addToast(t("toast.banChecking"), { durationMs: null });
-      checkingToastId = activeBanCheckToastId;
-    }
-
-    console.info("[ban-check] started", {
-      totalAccounts: steamIds.length,
-      idsToFetch: idsToFetch.length,
-      delayDays,
-      forceBanRefresh,
-      withinDelayWindow,
-    });
-
-    try {
-      const bans = await getPlayerBans(idsToFetch);
-      let bannedCount = 0;
-      const returnedIds = new Set<string>();
-      let malformedRows = 0;
-      for (const ban of bans) {
-        if (typeof ban.steam_id !== "string" || ban.steam_id.length === 0) {
-          malformedRows++;
-          continue;
-        }
-        banStates[ban.steam_id] = ban;
-        returnedIds.add(ban.steam_id);
-        if (ban.vac_banned || ban.community_banned || ban.number_of_game_bans > 0) {
-          bannedCount++;
-        }
-      }
-      if (malformedRows > 0) {
-        console.error("[ban-check] malformed ban rows without steam_id", {
-          malformedRows,
-          totalRows: bans.length,
-          sampleRow: bans[0] ?? null,
-        });
-      }
-      for (const steamId of idsToFetch) {
-        sessionBanCheckedIds.add(steamId);
-      }
-      if (returnedIds.size !== idsToFetch.length) {
-        const missingIds = idsToFetch.filter(steamId => !returnedIds.has(steamId));
-        console.warn("[ban-check] missing results for some Steam IDs", {
-          expected: idsToFetch.length,
-          received: returnedIds.size,
-          missingIds,
-        });
-      }
-
-      if (delayDays > 0) {
-        const mergedCheckedIds = forceBanRefresh || !withinDelayWindow
-          ? steamIds
-          : Array.from(new Set([...(cachedState?.checkedSteamIds ?? []), ...idsToFetch]));
-        writeBanCheckState({
-          lastSuccessAt: now,
-          checkedSteamIds: mergedCheckedIds,
-        });
-      } else {
-        localStorage.removeItem(BAN_CHECK_STATE_KEY);
-      }
-      writeBanInfoCache(Object.fromEntries(Object.entries(banStates)));
-
-      if (!silent && bannedCount > 0) {
-        addToast(
-          t(bannedCount > 1 ? "toast.banCheckSummary.multiple" : "toast.banCheckSummary.single", {
-            count: bannedCount,
-          })
-        );
-      }
-
-      console.info("[ban-check] completed", {
-        checkedAccounts: idsToFetch.length,
-        returnedRows: bans.length,
-        bannedCount,
-      });
-    } catch (e) {
-      if (!silent && now - lastBanErrorToastAt >= BAN_ERROR_TOAST_COOLDOWN_MS) {
-        addToast(t("toast.banCheckFailed", { error: String(e) }));
-        lastBanErrorToastAt = now;
-      }
-      console.error("[ban-check] failed to fetch ban states:", e);
-    } finally {
-      if (checkingToastId && activeBanCheckToastId === checkingToastId) {
-        removeToast(checkingToastId);
-        activeBanCheckToastId = null;
-      }
-    }
+  async function loadWarningStatesForAccounts(
+    adapter: PlatformAdapter,
+    accts: PlatformAccount[],
+    silent = true,
+    forceRefresh = false,
+  ) {
+    if (!adapter.loadWarningStates || accts.length === 0) return;
+    warningStates = await adapter.loadWarningStates(accts, { forceRefresh, silent, t });
   }
 
   async function load(
@@ -342,6 +138,7 @@ export function createAccountLoader(
     loading = true;
     error = null;
     try {
+      warningStates = adapter.getCachedWarningStates?.({ t }) ?? {};
       if (adapter.getStartupSnapshot) {
         const snapshot = await adapter.getStartupSnapshot();
         accounts = snapshot.accounts;
@@ -350,10 +147,11 @@ export function createAccountLoader(
         accounts = await adapter.loadAccounts();
         currentAccount = await adapter.getCurrentAccount();
       }
-      if (getActiveTab() === "steam" && accounts.length === 0) {
+      if (accounts.length === 0) {
         const now = Date.now();
-        if (now - lastNoAccountsToastAt >= LOAD_TOAST_COOLDOWN_MS) {
-          addToast(t("toast.noSteamAccountsFound"));
+        const emptyToast = adapter.getNoAccountsToastMessage?.({ t });
+        if (emptyToast && now - lastNoAccountsToastAt >= LOAD_TOAST_COOLDOWN_MS) {
+          addToast(emptyToast);
           lastNoAccountsToastAt = now;
         }
       } else if (showRefreshedToast && !silent) {
@@ -379,10 +177,12 @@ export function createAccountLoader(
       error = message;
       accounts = [];
       currentAccount = "";
-      if (getActiveTab() === "steam" && isSteamPathMissingError(message)) {
+      warningStates = {};
+      const errorToast = adapter.getLoadErrorToastMessage?.(message, { t });
+      if (errorToast) {
         const now = Date.now();
         if (now - lastSteamPathToastAt >= LOAD_TOAST_COOLDOWN_MS) {
-          addToast(t("toast.steamFolderNotFound"));
+          addToast(errorToast);
           lastSteamPathToastAt = now;
         }
       }
@@ -439,14 +239,15 @@ export function createAccountLoader(
     silent = true,
     deferBackground = true,
   ): Promise<number> {
+    const adapter = getAdapter();
     const visibleAccounts = resolveVisibleAccounts(accounts);
     if (visibleAccounts.length === 0) return 0;
     const run = async () => {
       const tasks: Promise<unknown>[] = [
         loadProfilesForAccounts(visibleAccounts, forceRefresh),
       ];
-      if (checkBans) {
-        tasks.push(fetchBanStates(visibleAccounts, silent, forceRefresh));
+      if (checkBans && adapter?.loadWarningStates) {
+        tasks.push(loadWarningStatesForAccounts(adapter, visibleAccounts, silent, forceRefresh));
       }
       await Promise.all(tasks);
       return visibleAccounts.length;
@@ -470,7 +271,7 @@ export function createAccountLoader(
     get switching() { return switching; },
     get error() { return error; },
     get avatarStates() { return avatarStates; },
-    get banStates() { return banStates; },
+    get warningStates() { return warningStates; },
     load,
     switchTo,
     addNew,
